Introduce logging functions that check ${VERBOSE}
Use of logging functions in initscripts are often conditioned on value
of $VERBOSE variable, causing same code to be repeated over and over
again.

This change introduce number of logging functions (with prefix "v") that
are functionally equivalent to their counterparts without prefix when
value of $VERBOSE is not "no", and do nothing if value of $VERBOSE is "no".

Closes: #928920