Fix the return value of vlog* functions when VERBOSE=no
To improve compatibility between the logging functions and the
vlog* ones, the return value should be consistent with their
counterparts, regardless of the VERBOSE value.

Each function will return 0 if VERBOSE=no, with the exception
of vlog_end_msg that will return the value of its first
argument, or 1 if not specified.